Move getIsExtensionAvailable to utils
