+++ src/utils/platform/getIsExtensionAvailable.ts
