+++ src/utils/platform/index.ts
@@ -1,0 +1,2 @@
+export * from './detectCurrentPlatform';
+export * from './getIsExtensionAvailable';
